refactor(competition): change u8 to u16 and increase the range of values to avoid overflow

+++ crates/core/src/competition/value_objects/league/row.rs
@@ -3,11 +3,11 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct LeagueTableRow {
     pub team_id: u64,
-    pub played: u8,
-    pub wins: u8,
-    pub draws: u8,
-    pub losses: u8,
-    pub goals_scored: u8,
-    pub goals_conceded: u8,
-    pub points: u8,
+    pub played: u32,
+    pub wins: u32,
+    pub draws: u32,
+    pub losses: u32,
+    pub goals_scored: u32,
+    pub goals_conceded: u32,
+    pub points: u32,
 }
